+++ src/Request/Files/Uploaded.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Nip\Request\Files;
-
-class Uploaded
-{
-    //@todo IMPLREMENTATION
-}