+++ tests/unit/RequestTest.php
@@ -3,50 +3,51 @@
 namespace Nip\Tests\Unit;
 
 use Nip\Request;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
 
-class RequestTest extends \Codeception\TestCase\Test
+/**
+ * Class RequestTest
+ * @package Nip\Tests\Unit
+ */
+class RequestTest extends AbstractTest
 {
-    /**
-     * @var \UnitTester
-     */
-    protected $tester;
 
     /**
      * @var \Nip\Request
      */
-    protected $_object;
+    protected $request;
 
     public function testAttributeSet()
     {
-        $this->_object->setAttribute('bar','foo');
+        $this->request->setAttribute('bar', 'foo');
 
-        static::assertEquals('foo', $this->_object->attributes->get('bar'));
-        static::assertEquals('foo', $this->_object->get('bar'));
+        static::assertEquals('foo', $this->request->attributes->get('bar'));
+        static::assertEquals('foo', $this->request->get('bar'));
     }
 
     public function testQuerySet()
     {
-        $this->_object->query->set('bar','foo');
+        $this->request->query->set('bar', 'foo');
 
-        static::assertEquals('foo', $this->_object->query->get('bar'));
-        static::assertEquals('foo', $this->_object->get('bar'));
+        static::assertEquals('foo', $this->request->query->get('bar'));
+        static::assertEquals('foo', $this->request->get('bar'));
     }
 
     // tests
 
     public function testBodySet()
     {
-        $this->_object->body->set('bar','foo');
+        $this->request->request->set('bar', 'foo');
 
-        static::assertEquals('foo', $this->_object->body->get('bar'));
-        static::assertEquals('foo', $this->_object->get('bar'));
+        static::assertEquals('foo', $this->request->request->get('bar'));
+        static::assertEquals('foo', $this->request->get('bar'));
     }
 
     public function testGetOrder()
     {
         $this->testInitialize();
 
-        static::assertEquals('value44', $this->_object->get('var4'));
+        static::assertEquals('value44', $this->request->get('var4'));
     }
 
     public function testInitialize()
@@ -57,11 +58,11 @@ class RequestTest extends \Codeception\TestCase\Test
         $post['var4'] = 'value4';
         $attributes['var4'] = 'value44';
 
-        $this->_object->initialize($get, $post, $attributes);
+        $this->request->initialize($get, $post, $attributes);
 
-        static::assertEquals($get, $this->_object->query->all());
-        static::assertEquals($post, $this->_object->body->all());
-        static::assertEquals($attributes, $this->_object->attributes->all());
+        static::assertEquals($get, $this->request->query->all());
+        static::assertEquals($post, $this->request->request->all());
+        static::assertEquals($attributes, $this->request->attributes->all());
     }
 
     public function testCreateFromGlobals()
@@ -69,15 +70,15 @@ class RequestTest extends \Codeception\TestCase\Test
         $_GET['foo1'] = 'bar1';
         $_POST['foo2'] = 'bar2';
         $_COOKIE['foo3'] = 'bar3';
-        $_FILES['foo4'] = array('bar4');
+        $_FILES['foo4'] = ['bar4'];
         $_SERVER['foo5'] = 'bar5';
 
         $request = Request::createFromGlobals();
 
         static::assertEquals('bar1', $request->query->get('foo1'), '::fromGlobals() uses values from $_GET');
-        static::assertEquals('bar2', $request->body->get('foo2'), '::fromGlobals() uses values from $_POST');
+        static::assertEquals('bar2', $request->request->get('foo2'), '::fromGlobals() uses values from $_POST');
         static::assertEquals('bar3', $request->cookies->get('foo3'), '::fromGlobals() uses values from $_COOKIE');
-        static::assertEquals(array('bar4'), $request->files->get('foo4'), '::fromGlobals() uses values from $_FILES');
+        static::assertEquals(['bar4'], $request->files->get('foo4'), '::fromGlobals() uses values from $_FILES');
         static::assertEquals('bar5', $request->server->get('foo5'), '::fromGlobals() uses values from $_SERVER');
     }
 
@@ -87,8 +88,8 @@ class RequestTest extends \Codeception\TestCase\Test
         $request->setActionName('action1');
         $request->setControllerName('controller1');
         $request->setModuleName('module1');
-        $atributes = array('attrb1' => 'val1', 'attrb2' =>'val2');
-        $request->attributes->add($atributes);
+        $attributes = ['attrb1' => 'val1', 'attrb2' => 'val2'];
+        $request->attributes->add($attributes);
 
         $duplicateAction = $request->duplicateWithParams('action2');
         static::assertEquals('action2', $duplicateAction->getActionName());
@@ -108,36 +109,34 @@ class RequestTest extends \Codeception\TestCase\Test
 
     public function testCreateFromGlobalsFiles()
     {
-        $_FILES = array(
-            'file1' => array(
+        $_FILES = [
+            'file1' => [
                 'name' => 'MyFile.txt',
                 'type' => 'text/plain',
-                'tmp_name' => '/tmp/php/php1h4j1o',
+                'tmp_name' => codecept_data_dir('Request/php1h4j1o'),
                 'error' => UPLOAD_ERR_OK,
                 'size' => 123,
-            ),
-            'file2' => array(
+            ],
+            'file2' => [
                 'name' => 'MyFile.jpg',
                 'type' => 'image/jpeg',
-                'tmp_name' => '/tmp/php/php1h4j1o',
+                'tmp_name' => codecept_data_dir('Request/php1h4j1o'),
                 'error' => UPLOAD_ERR_OK,
                 'size' => 300,
-            )
-        );
+            ],
+        ];
 
         $request = Request::createFromGlobals();
 
-        $this->isInstanceOf('Nip\Request\Files\Uploaded', $request->files->get('file2'), '::fromGlobals() uses values from $_FILES');
+        static::assertInstanceOf(
+            UploadedFile::class,
+            $request->files->get('file2'),
+            '::fromGlobals() uses values from $_FILES');
     }
 
-    protected function _before()
+    protected function setUp()
     {
-        $this->_object = new Request();
+        parent::setUp();
+        $this->request = new Request();
     }
-
-    protected function _after()
-    {
-    }
-    
-    
 }